add plugin for connections by chart type
