+++ src/ui/registry/units/ql/functions-map.ts
@@ -1,5 +1,8 @@
+import {QLChartType} from 'shared';
+
 import {makeFunctionTemplate} from '../../../../shared/utils/makeFunctionTemplate';
 
 export const qlFunctionsMap = {
     getDefaultMonitoringQLConnectionId: makeFunctionTemplate<(env: string) => string>(),
+    getConnectionsByChartType: makeFunctionTemplate<(chartType: QLChartType) => string[]>(),
 } as const;
